Extract Vehicle Search form to own Component
- MiniVehicleSearchForm extracted from header
  - Form registered in main components
- Component runs wil useHistory instead of setState() for applying query
