+++ client/src/components/index.js
@@ -4,6 +4,7 @@ export { default as Safety } from "./analytics/Safety";
 
 // Find
 export { default as SearchResults } from "./find_cars/SearchResults";
+export { default as MiniVehicleSearchForm } from "./find_cars/MiniVehicleSearchForm";
 export { default as FindByMake } from "./find_cars/FindByMake";
 export { default as FindByModel } from "./find_cars/FindByModel";
 export { default as FindByVin } from "./find_cars/FindByVin";
